+++ Tests/bin/runtests.sh
@@ -6,10 +6,14 @@ source $(dirname ${BASH_SOURCE[0]})/set-env-vars.sh
 
 VERBOSITY=
 RESET=false
+COVERAGE=
 
-while getopts "vr" opt
+while getopts ":c:vr" opt
 do
     case $opt in
+        c)
+            COVERAGE="--coverage-clover=${OPTARG}"
+        ;;
         r)
             RESET=true
         ;;
@@ -27,4 +31,6 @@ fi
 
 # Note: make sure we run the version of phpunit we installed, not the system one. See: https://github.com/sebastianbergmann/phpunit/issues/2014
 
-$(dirname $(dirname $(dirname ${BASH_SOURCE[0]})))/vendor/phpunit/phpunit/phpunit --stderr --colors ${VERBOSITY} Tests/phpunit "$@"
+# @todo pass to phpunit any further cli options, or at least allow to specify a single test using --filter
+
+$(dirname $(dirname $(dirname ${BASH_SOURCE[0]})))/vendor/phpunit/phpunit/phpunit --stderr --colors ${VERBOSITY} ${COVERAGE} Tests/phpunit "$@"