+++ WHATSNEW.md
@@ -29,7 +29,8 @@ Version 5.13.0
 
 * New: taught the `kaliop:migration:status` command to sort migrations by execution date using `--sort-by` (issue #224)
 
-* New: taught the test-execution command `teststack.sh` a new action: `console`, as well as new option `-r runtests`
+* New: taught the test-execution command `teststack.sh` two new actions: `console` and `dbconsole`, as well as a new
+  option: `-r runtests`
 
 * Improved: a single value for a content field of type ezcountry can be specified as a string instead of an array (issue #190)